Make startParanetSync schedule paranetSync commands for each paranet
